==> 0. Explore topics.ipynb <==
%load_ext autoreload
%autoreload 2

from time import time
import pandas as pd
import numpy as np
import os
import json

topics_path = "/data3/crisIda_Shared/CAST_2022/CAST_2022_topics/2022_evaluation_topics_tree_v1.0.json"

topics_df = pd.read_json(topics_path)

for i in range(len(topics_df)):
    print(i)
    conv_tree_list = topics_df["turn"][i]
    for turn in conv_tree_list:
        print(turn)
        if turn["participant"] == "User":
            if "parent" in turn:
                print("$$$$$$$$$", turn["number"], turn["parent"], turn["manual_rewritten_utterance"])
            else:
                print("$$$$$$$$$", turn["number"], turn["manual_rewritten_utterance"])
        print()
    break

## All conversation paths for each conversation for AUTO and MANUAL

auto_topics_path = "/data3/crisIda_Shared/2022_automatic_evaluation_topics_flattened_duplicated_v1.0.json"

man_topics_path = "/data3/crisIda_Shared/CAST_2022/CAST_2022_topics/2022_evaluation_topics_flattened_duplicated_v1.0.json"

with open("/data3/muntean/TREC_CAST_2022/data/raw_conv.tsv", "w") as output:
    with open(auto_topics_path) as topics_file:
            topics = json.load(topics_file)
            for topic in topics:
                for idx, turn in enumerate(topic['turn']):
                    turn_number = f"{topic['number']}_{turn['number']}"
                    # utterance = turn.get('automatic_rewritten_utterance')
                    # utterance = turn.get('manual_rewritten_utterance')
                    utterance = turn.get('utterance')
                    print(turn_number, utterance)
                    output.write(turn_number + "\t" + utterance + "\n")
                print()
                output.write("\n")
                

## Manual rewritten

man_topics_path = "/data3/crisIda_Shared/CAST_2022/CAST_2022_topics/2022_evaluation_topics_tree_v1.0.json"
manual_utt_path_drhard = "/data3/muntean/DRhard/data/passage/dataset/CAST-2022/queries.CAST2022manual.test.tsv"

with open(manual_utt_path_drhard, 'w') as output:
    with open(man_topics_path) as topics_file:
            topics = json.load(topics_file)
            for topic in topics:
                for idx, turn in enumerate(topic['turn']):
                    turn_number = f"{topic['number']}_{turn['number']}"
                    # utterance = turn.get('automatic_rewritten_utterance')
                    utterance = turn.get('manual_rewritten_utterance')
                    if utterance is not None:
                        print(turn_number, utterance)
                        output.write('\t'.join([turn_number, utterance])+'\n')
                print()

## Automatic rewritten

man_topics_path = "/data3/crisIda_Shared/CAST_2022/CAST_2022_topics/2022_automatic_evaluation_topics_tree_v1.0.json"
manual_utt_path_drhard = "/data3/muntean/DRhard/data/passage/dataset/CAST-2022/queries.CAST2022automatic.test.tsv"

with open(manual_utt_path_drhard, 'w') as output:
    with open(man_topics_path) as topics_file:
            topics = json.load(topics_file)
            for topic in topics:
                for idx, turn in enumerate(topic['turn']):
                    turn_number = f"{topic['number']}_{turn['number']}"
                    # utterance = turn.get('automatic_rewritten_utterance')
                    utterance = turn.get('automatic_rewritten_utterance')
                    if utterance is not None:
                        print(turn_number, utterance)
                        output.write('\t'.join([turn_number, utterance])+'\n')
                print()
